+++ nuxeo-drive-client/tests/test_local_changes_when_offline.py
@@ -5,15 +5,15 @@ later when Drive becomes online.
 """
 
 import shutil
+import sys
 
-from nxdrive.osi import AbstractOSIntegration
 from tests.common_unit_test import FILE_CONTENT, UnitTestCase
 
-if AbstractOSIntegration.is_windows():
+if sys.platform == 'win32':
     from win32com.shell import shell, shellcon
 else:
     import xattr
-    if AbstractOSIntegration.is_mac():
+    if sys.platform == 'darwin':
         import Cocoa
 
 
@@ -48,11 +48,11 @@ class TestOfflineChangesSync(UnitTestCase):
         shutil.copytree functionality.
         """
 
-        if AbstractOSIntegration.is_windows():
+        if sys.platform == 'win32':
             # Make a copy of file1 using shell (to copy including xattr)
             shell.SHFileOperation((0, shellcon.FO_COPY, src, dst,
                                    shellcon.FOF_NOCONFIRMATION, None, None))
-        elif AbstractOSIntegration.is_mac():
+        elif sys.platform == 'darwin':
             fm = Cocoa.NSFileManager.defaultManager()
             fm.copyItemAtPath_toPath_error_(src, dst, None)
         else: