+++ nuxeo-drive-client/tests/test_multiple_files.py
@@ -1,10 +1,11 @@
 # coding: utf-8
 import os
 import shutil
-from unittest import skipIf
+import sys
+
+import pytest
 
 from nxdrive.logging_config import get_logger
-from nxdrive.osi import AbstractOSIntegration
 from tests.common_unit_test import FILE_CONTENT, RandomBug, UnitTestCase
 
 log = get_logger(__name__)
@@ -143,9 +144,10 @@ class MultipleFilesTestCase(UnitTestCase):
                                          os.path.dirname(self.folder_path_1),
                                          stopped=False)
 
-    @skipIf(AbstractOSIntegration.is_linux(),
-            'NXDRIVE-471: Not handled under GNU/Linux as'
-            ' creation time is not stored')
+    @pytest.mark.skipif(
+        sys.platform == 'linux2',
+        reason='NXDRIVE-471: Not handled under GNU/Linux'
+               ' as creation time is not stored.')
     def test_move_and_copy_paste_folder_original_location_stopped(self):
         self._move_and_copy_paste_folder(self.folder_path_1, self.folder_path_2,
                                          os.path.dirname(self.folder_path_1))