+++ docs/changes.md
@@ -23,6 +23,7 @@ Release date: `2017-??-??`
 - GUI: Fix a SEGFAULT when closing the metadata window
 - GUI: Add envar `USE_OLD_MENU` to force the use of the old menu
 - Jenkins: The beta job now uploads the package to the PyPi server
+- Tests: Big clean-up
 - \[macOS\] Fix the favorite link creation
 - \[Windows\] Packaging: Prevent infinite loop when uninstalling
 - \[Windows\] Packaging: Fix symbolic link creation